# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MotoTrack ERP/POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers for current company
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vouchers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
    post:
      operationId: createVoucher
      tags: [vouchers]
      summary: Create a voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
  /vouchers/{id}:
    get:
      operationId: getVoucher
      tags: [vouchers]
      summary: Get voucher with entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voucher detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetail"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role, active]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
        role:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
        active:
          type: boolean
        password:
          type: string

    # Location
...
    Voucher:
      type: object
      required: [id, companyId, voucherNumber, voucherType, voucherDate]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        voucherNumber:
          type: string
        voucherType:
          type: string
        voucherDate:
          type: string
        description:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        createdAt:
          type: string
    VoucherInput:
      type: object
      required: [voucherType, voucherDate]
      properties:
        voucherType:
          type: string
        voucherDate:
          type: string
        description:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VoucherEntryInput"
    VoucherEntryInput:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: integer
        debitAmount:
          type: number
        creditAmount:
          type: number
        narration:
          type: string
    VoucherEntry:
      type: object
      required: [id, voucherId, accountId]
      properties:
        id:
          type: integer
        voucherId:
          type: integer
        accountId:
          type: integer
        accountName:
          type: ["string", "null"]
        debitAmount:
          type: ["number", "null"]
        creditAmount:
          type: ["number", "null"]
        narration:
          type: ["string", "null"]
    VoucherDetail:
      type: object
      required: [id, companyId, voucherNumber, voucherType, voucherDate]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        voucherNumber:
          type: string
        voucherType:
          type: string
        voucherDate:
          type: string
        description:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        createdAt:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VoucherEntry"

    # Employee
...