# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MotoTrack ERP/POS API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  role: string;
  active: boolean;
  createdAt?: string;
}