# lib/api-zod/src/generated/types/voucherEntry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MotoTrack ERP/POS API
 * OpenAPI spec version: 0.1.0
 */

export interface VoucherEntry {
  id: number;
  voucherId: number;
  accountId: number;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  debitAmount?: number | null;
  /** @nullable */
  creditAmount?: number | null;
  /** @nullable */
  narration?: string | null;
}